requirements.txt
alembic==1.13.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
config==0.5.1
Flask==3.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.0.0
opencv-python==4.10.0.84
requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
